refactor(bbapi/avm): inline execute() into handlers, delete domain AVM command structs
POC for the larger handler-inlining refactor. The 3 AVM commands
(AvmProve, AvmVerify, AvmCheckCircuit) had domain structs that mirrored
their wire counterparts exactly except for `proof: vector<bb::fr>` vs
`vector<Fr>`. The handlers were converting wire→domain, calling
execute(), then converting back.

Inline the bodies of `AvmProve::execute()`,`AvmVerify::execute()`, and
`AvmCheckCircuit::execute()` directly into the corresponding
`handle_avm_<method>(BBApiRequest&, wire::Cmd&&)` functions. Field-element
conversion happens once at the boundary (`fr_vec_from_wire(cmd.proof)`)
rather than on the whole struct. The AVM stats helpers
(`reset_avm_stats` / `snapshot_avm_stats_wire`) move into the handlers
TU's anonymous namespace and now produce `wire::AvmStat` directly.

- Delete `bbapi_avm.{hpp,cpp}` — the only AVM domain types lived here.
- Remove AvmProve / AvmVerify / AvmCheckCircuit from the
  `bbapi::Command` / `CommandResponse` NamedUnions in
  `bbapi_execute.hpp`. AVM is exclusively reachable via `handle_avm_*`
  now.
- Drop `avm_stat_{to_wire,vec_to_wire}` from `bbapi_wire_convert.hpp`.

bbapi_tests 30/30, schema dump unchanged (wire types emit the same
NamedUnion shape via the codegen-emitted reflection).

Crypto / ecc / ecdsa / schnorr / srs / ultra_honk / chonk get the same
treatment in follow-up commits.
